# HeroModule/hero-service/src/main/java/ru/itmo/cs/parsifal/heroservice/model/Coordinates.java
package ru.itmo.cs.parsifal.heroservice.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Coordinates {
    private Integer x;
    private Double y;
}